// src/models/GameChar.ts
export enum GameChar {
    W_ROOKS_CHAR = 'r',
    W_KNIGHTS_CHAR = 'n',
    W_BISHOPS_CHAR = 'b',
    W_QUEENS_CHAR = 'q',
    W_KING_CHAR = 'k',
    W_PAWNS_CHAR = 'p',
    B_ROOKS_CHAR = 'R',
    B_KNIGHTS_CHAR = 'N',
    B_BISHOPS_CHAR = 'B',
    B_QUEENS_CHAR = 'Q',
    B_KING_CHAR = 'K',
    B_PAWNS_CHAR = 'P'
}